cuda-evm/eternal-keras/eai/__init__.py
from eai.func import transform, check, get_model
from eai.version import __version__
import os
from dotenv import load_dotenv
from eai.utils import ENV_PATH, publisher
os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
load_dotenv(ENV_PATH)
